Move database repo implementations to database package
